requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
idna==2.8
praw==6.1.1
prawcore==1.0.1
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
update-checker==0.16
urllib3==1.24.1
websocket-client==0.54.0
